requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
celery==5.3.4
redis==5.0.1
httpx==0.25.2
beautifulsoup4==4.12.2
google-generativeai==0.3.1
numpy==1.26.2
scikit-learn==1.3.2
python-dotenv==1.0.0
alembic==1.13.0
